# apps/backend/swagger.yml
openapi: 3.1.0
info:
  title: POOSD - Group 1
  contact:
    url: https://github.com/UCF-GroupProj/poosd-monorepo
  version: '1.0.0'
servers:
  - url: http://localhost:8080
  - url: https://api.poosd.zhiyan114.com
tags:
  - name: Auth
    description: Credential Management Endpoint
  - name: Password Reset
    description: Password Reset Endpoint
  - name: Email Verification
    description: Email Verification Endpoint
  - name: Card Info
    description: Get Information about user's card collections
  - name: User Profile
    description: Get Information about non-card related user stats
paths:
  /login:
    post:
      tags:
        - Auth
      summary: Log in a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token 
        '403':
          description: Account Not Enabled
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Server error
          content:
            text/plain:
              schema:
                type: string
  /register:
    post:
      tags:
        - Auth
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: User registered successfully
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Malformed Body
          content:
            text/plain:
              schema:
                type: string
        '409':
          description: Duplicate Records (such as email)
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Server error
          content:
            text/plain:
              schema:
                type: string
  /pwdreset/{ResetCode}:
    parameters:
      - name: ResetCode
        in: path
        required: true
        schema:
          type: string
        description: Password reset code
    get:
      tags:
        - Password Reset
      summary: Get password reset status or info
      responses:
        '200':
          description: Password reset information retrieved successfully
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "{userID}"
        '404':
          description: Provided Reset Code not found
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "No valid password request found with this code"
        '500':
          description: Server error
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Server error"
        '503':
          description: Service unavailable
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Database is experiencing issues"
    patch:
      tags:
        - Password Reset
      summary: Complete password reset by updating the password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newPassword
              properties:
                newPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Password reset successfully
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Password Successfully Updated!"
        '400':
          description: Malformed Body or invalid token
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Invalid/Expired Request Token
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Server error
          content:
            text/plain:
              schema:
                type: string
  /pwdreset:
    post:
      tags:
        - Password Reset
      summary: Initiate a password reset request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Password Reset Request Submitted
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Verification email has been sent, if there's an account associated with this email"
        '400':
          description: Invalid Request
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Missing Email Field"
        '500':
          description: Server error
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Server error"
        '503':
          description: Service unavailable
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Database is experiencing issues"
  /verifyemail/{verifyCode}:
    parameters:
      - name: verifyCode
        in: path
        required: true
        schema:
          type: string
        description: Email Verification Code
    get:
      tags:
        - Email Verification
      summary: Get email verification status or info
      responses:
        '200':
          description: Email verification information retrieved successfully
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "{userID}"
        '404':
          description: verifyCode invalid
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Provided verification code was not found"
        '500':
          description: Server error
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Generic backend error"
    patch:
      tags:
        - Email Verification
      summary: Complete email verification by updating the verification status
      responses:
        '200':
          description: Email verified successfully
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Email Successfully Verified"
        '404':
          description: Verification Code Doesn't exist
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Verification code isn't valid"
        '500':
          description: Server error
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Generic Server Error"
        '503':
          description: Service Unavailable
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Database Communication Error"
  /card/{cardID}:
    get:
      security:
        - BearerAuth: []
      tags:
        - Card Info
      summary: Get information about a specific card by ID
      parameters:
        - name: cardID
          in: path
          required: true
          schema:
            type: string
          description: Card identifier
      responses:
        '200':
          description: Card information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  description:
                    type: string
                  owned:
                    type: boolean
                  imgURI:
                    type: string
                description: Card information object
        '401':
          description: Bad Authorization Token
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Bad Authorization Token"
        '404':
          description: Card not found
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Server error
          content:
            text/plain:
              schema:
                type: string
  /card/summary:
    get:
      security:
        - BearerAuth: []
      tags:
        - Card Info
      summary: Get summary information about user's card collections
      responses:
        '200':
          description: Card collection summary retrieved successfully
          content:
            application/json:
              schema:
                type: object
                description: Summary of user's card collections
                properties:
                  totalUniqueCards:
                    type: integer
                  commonOwned:
                    type: integer
                  rareOwned:
                    type: integer
                  epicOwned:
                    type: integer
                  legendaryOwned:
                    type: integer
        '401':
          description: Bad Authorization Token
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Bad Authorization Token"
        '500':
          description: Server error
          content:
            text/plain:
              schema:
                type: string
  /profile/:
    get:
      tags:
        - User Profile
      security:
        - BearerAuth: []
      summary: Get user account's stats (non-card related)
      responses:
        '200':
          description: User profile information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                description: User profile information
                properties:
                  id:
                    type: string
                  verified:
                    type: boolean
                  collection:
                    type: array
                    items:
                      type: string
                      description: ObjectId
                  level:
                    type: integer
                  exp:
                    type: integer
                  currency:
                    type: object
                    properties:
                      gems:
                        type: integer
                  favorites:
                    type: array
                    items:
                      type: string
                      description: ObjectId
                  pullsSinceEpic:
                    type: integer
                  lastPullTime:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "Bad Authorization Token"
        '500':
          description: Server error
          content:
            text/plain:
              schema:
                type: string
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT